requirements.txt
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
backcall==0.2.0
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
daphne==3.0.2
decorator==5.1.1
dj-database-url==0.5.0
Django==4.0
django-channels==0.7.0
django-rest-knox==4.2.0
django-stubs==1.12.0
django-stubs-ext==0.5.0
djangorestframework==3.13.1
executing==0.8.3
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
ipython==8.4.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mypy==0.971
mypy-extensions==0.4.3
oauthlib==3.2.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
pyOpenSSL==22.0.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
tomli==2.0.1
traitlets==5.3.0
Twisted==22.4.0
txaio==22.2.1
types-pytz==2022.1.2
types-PyYAML==6.0.10
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.10
wcwidth==0.2.5
zope.interface==5.4.0
